Move common python code to root and create a setup.py
