+++ funky_lights/crc16.py
