+++ requirements.txt
@@ -1,0 +1,11 @@
+adsk==2.0.12157
+crc8==0.1.0
+intelhex==2.3.0
+matplotlib==3.5.1
+numpy==1.21.5
+opencv_python==4.5.5.64
+pyserial==3.5
+pyserial_asyncio==0.6
+setuptools==61.2.0
+SQLAlchemy==1.4.32
+websockets==10.3
